requirements.txt
Django
djangorestframework
django-filter
Markdown
Pillow
